requirements.txt
absl-py
asn1crypto
boto
beautifulsoup4
cachetools
certifi
cffi
chardet
cryptography
decorator
discovery
enum34
flask
futures
gcs-oauth2-boto-plugin
gcsfs
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
google-cloud-bigquery
google-cloud-core
google-cloud-firestore
google-cloud-pubsub
google-cloud-storage
google-reauth
google-resumable-media
googleapis-common-protos
grpcio
httplib2
idna
ipaddress
lxml
messytables
oauth2client
oauthlib
protobuf
pyasn1
pyasn1-modules
pycparser
pyOpenSSL
pytz
pyu2f
requests
requests-oauthlib
retry-decorator
rsa
six
SocksiPy-branch
uritemplate
urllib3